Extract fetching logic for templates to it's own file
